Upload coverage report with different artifact names in CI
When the CI fails, we upload the coverage report as an artifact.

Previously, `actions/upload-artifact` allowed uploading artifacts with
the same name in different jobs in v3, but v4 doesn't allow this [^1].
We need to upload artifacts with different names; otherwise, it fails
with a 409 error. It seems we overlooked this change when upgrading from
v3 to v4.

[^1]: https://github.com/actions/upload-artifact?tab=readme-ov-file#not-uploading-to-the-same-artifact