Add disk monitoring in VM host
This commit contains changes for disk monitoring for both RAID and
non-RAID systems, supporting NVME and SCSI storage controller
interfaces. The system is designed to identify and report disk failures
in a pull-based monitoring setup.

Storage Device Model

To identify which disks are down in a pull-based system, we first need
to maintain the record of existing disks. We introduced a column in the
`storage_device` model to store the list of physical disks mapped to
each logical disk.

Note: Entries in the storage_device table do not represent physical
storage. For example, a RAID1 setup would only have a single entry in
the table despite being backed by multiple physical storage devices. Due
to this, the newly added column is a collection, not a single entry.

Device Identification

We decided to store the device's serial number rather than using more
legible device file naming (e.g., /dev/nvme0n1). This choice was made
due to the non-permanent nature of device file naming. Device names tend
to change on reboot or when ports on which they are attached to the
machine are physically swapped.

Mountpoint and Physical Device Association

To find the associated physical device for each storage_device table
entry, we leverage mountpoint information. The mountpoint can be
inferred from the name column of storage_device:
- If the name is `DEFAULT`, the disk is mounted at `/`.
- Otherwise, it is mounted at `/var/storage/devices/<name>`.

Precedence for this can be found [in this
commit](https://github.com/ubicloud/ubicloud/commit/89f9dbd7f2b12c08eb993d947e64003c68c90606#diff-f336d7f0af53644f9dae8673dcebdaab01a2d3173a00ba648fc48d79aee2ba81R19).

After finding the associated mountpoint, we parse lsblk to get the
serial number of the physical device.

Health Check Process

On each check pulse run: 1. We fetch the device file name for each
serial number.  2. We run a health check using smartmontools.  3. If any
device fails, it fails the overall `check_pulse`.

We supplement smartmontools with another check to detect if the OS has
remounted the disk as read-only. This occurs when the OS observes too
many IO failures, among other reasons. To check if the device is
read-only, we rely on the `lsblk` command.

Smartmontools doesn't cache the health check result. It works for both
NVME and SCSI, with differences in output formatting:
- For NVME: We see "PASSED"
- For SCSI: We see "OK"

This is cross-checked from their source code on
[GitHub](https://github.com/smartmontools/smartmontools/blob/65707ca59285a4b2350fb5b30293398f17e868bd/smartmontools/scsiprint.cpp#L402).
Also, smartmontools are not part of linux distributions. It needs to be
installed. We do this in `rhizome/host/bin/prep_host.rb`.

Failure Modes

1. SSH fails:
   - If due to binary not found: The disk on which SSH resides gets
     detached. Systems are expected to survive such events.
   - If due to other reasons (e.g., machine busy or out of resources):
     We fail the check.

2. `smartctl` binary not found:
   - This could be due to disk detachment. We're currently ignoring this
     scenario.
   - Reasoning: The smartctl binary's page should be in memory since we
     run it every 30 seconds. However, we can't make assumptions about
client usage.
   - Potential solution: Pin the page to RAM using `vmtouch`.

3. `smartctl` command fails:
   - We fail the check in this case.

These checks are run per storage as part of the VmHost monitoring,
reusing the strand.

Other Considerations

- Populating serial numbers in the storage_device table:
  - Option 1: Manually run a script at new deployment time. This
    requires considering failure retry mechanisms and handling permanent
failures.
  - Option 2 (Chosen): Anchor population to the users of this column.
    When the storage_device model is asked for a serial number it
doesn't have, it runs an SSH command to find the serial number,
populates the table, and returns to the caller.

- Testing the mountpoint to serial number mapping:
  - We created loopback devices, which added more entries to the lsblk
    output.
  - Running the logic for fetching serial numbers from the control plane
    retrieved the expected serial numbers.
  - The script used to create loopback fs is [mentioned in this commit
    message](https://github.com/ubicloud/ubicloud/commit/89f9dbd7f2b12c08eb993d947e64003c68c90606#diff-f336d7f0af53644f9dae8673dcebdaab01a2d3173a00ba648fc48d79aee2ba81R19).

- LearnStorage update:
  - We updated LearnStorage to populate serial numbers in
    storage_device.
  - LearnStorage runs as part of the PrepHost program, which is executed
    on machine changes like new disk attachments.

- Disk CRUD:
  - Removal of disk and addition of new disk to existing RAID systems is
    still to be determined (TBD).
  - Addition of new disk is handled by `LearnStorage` prog.

- Transaction consideration:
  - A transaction is not required for populating serial numbers in
    storage_device as the race condition is benign.