Remove usage of class variables
Class variables are a Ruby misfeature that is almost never a good
idea to use.  Replace most usage with class instance variables, and
one case with a constant.

Add freeze methods to classes switching from class variables to
class instance variables, which ensure the class instance variable
is set before the class is frozen.