Fix 2 nondeterministic tests by hardcoding expected value
Without this change, I see failures such as:

```
Failure/Error: expect(page.title).to eq("Ubicloud - #{Account[email: TEST_USER_EMAIL].projects.first.name} Dashboard")

expected: "Ubicloud - Default Dashboard"
     got: "Ubicloud - Invited project Dashboard"
```

This is because there are multiple projects for the account,
and a dataset order was not enforced.  We could enforce a
dataset order, but since the value we are expecting is
constant, it seems simpler to hardcode it.