Change Routes::Common::VmHelper#post to Clover#vm_post
This removes Routes::Common::VmHelper as it is no longer used.

This completes the merging of the project/location/vm and
project/vm routes.  I think the merged routes are significantly
simpler to understand.  In most cases, indirection is avoided.
For the list and post routes, there is still indirection to
a helper method, but the helper method has been simplied, and
no longer requires a context switch, since it runs in the
same context as the routing tree.