Add bin/production_check for checking app loads in production
This file starts puma, and checks for correct loading or for
failure.  It returns 0 on success and 1 on failure, printing
the error to stderr in the failure case.

Run this before the control plane tests in CI.  I've tested
that this file catches the production issue in my deployment
this morning.