Avoid unnecessary model reloading in private submit routes
Neither variable is used after the reload, so I don't think the
reloads are unnecessary, unless you are trying to check for
concurrent deletion, and I don't think that is the reason.

In the connect route, there are actual two different reloads,
as both ps.reload and PrivateSubnet[ps.id] will cause a query.
Unless the path of a private subnet can change, using the
existing ps object seems better.