Change map(&:id).include?(id) to find
find can stop as soon as a matching entry is found, but map.include?
requires a full iteration, and then then same scan until a matching
entry is found.  This is inside of another loop, so it seems worth
optimizing this case.

Maybe we should have PrivateSubnet#connectable_subnets method for
this?