Add load balancer stack option
Our load balancers are applied to both IPv4 and IPv6 addresses of the
underlying VMs. This is not very useful for the customer in such a
scenario;
1. Customer creates and apps and uses only IPv4 to server the app.
2. Creates a load balancer and attaches his VMs.
3. We create the dns records for both IPv4 and IPv6 addresses of these
VMs for the load balancer.
4. Client of the customer resolves to the IPv6 address of the load
balancer but the server is not responsive in that address.

This is why we are adding stack option to the load balancer. For now,
this is only added as a backend option and will be made available soon
to the customers as well.