Move nic gateway setting to vm_setup rhizome changes
Nic gateway setup is the most crucial piece in ipv4 networking setup.
Most of that work already is happenning in vm_setup. Keeping this piece
in controlplane doesn't make sense. Therefore, I am moving it to the
proper place.

This will require migration of prep.json files, which is fine and will
be handled at the time of deployment for existing VMs.

This commit essentially adds the rhizome part, the deployment will
require to first deploy this, wait and continue the removal of gateway
setup from the control plane, which is the next commit.