Enforce 100% coverage when doing parallel coverage testing
Previously, serial coverage testing failed if coverage was
not 100%, but parallel coverage testing did not.  This is
because parallel coverage recomputes coverage in every child
process, and none of the child processes know which one will
be last.

This works around the issue by checking the coverage/index.html
file after the parallel coverage run, and failing if any lines
or branches were missed.