Use more consistent authorization in the project/location/firewall attach-subnet route
Previously, api was checking PrivateSubnet:edit against @project.id,
and web was checking Firewall:view against firewall.id

I'm not sure which is correct, but as specs pass with both
authorizations, this uses both.

Previously, web was also checking PrivateSubnet:edit against the
private_subnet.  I tried also doing that in api, but it results
in 2 failing specs.  This leaves the authorization as web-only,
but the differing authorization must be addressed and fixed
(probably by changing the api specs).

While here, DRY up and simplify the route. This route had the last
use of `user`, so remove the setting of that local variable.