Merge the project/location/firewall attach-subnet and detach-subnet routes
These have almost the same code, with the difference between only
in the method call (associate_with_private_subnet vs
disassociate_from_private_subnet), and flash message.  The flash
message was also inconsistent in whether the private subnet name
was mentioned, and this makes it consistent by including the name
in both cases.