Use Cache-Control:no-store for sensitive data
PostgreSQL get endpoint returns connection string with password in it. Clients
can cache the result and store in the client machine, which in turn can be
read by unauthorized people. We are setting `Cache-Control: no-store` to let
clients know that the response should not be cached. There is also another
header that we can set; `Pragma: no-cache`, however `Pragma` is deprecated and
it is only used by HTTP/1.0 clients, thus I don't think we need to set it.