Add flash-notice/flash-error id attributes
After we add these, we can update the specs to use them instead of
using `have_content`.  `have_content` checks for content anywhere
in the page, and the specs should be checking for error/notice
messages in a specific place.

This also makes it easier to debug failing specs, since you can
more easily see what the flash notice/error being displayed is.