Load pg_array extension before other extensions
As mentioned in the pg_json and pg_range documentation, you should
load pg_array first.  I doubt we are using json[], jsonb[], or
one of the array range types, but just in case we use array_agg
in a query with one of those types, we probably want to get an
array-like object back, and not a string.