Format mistakenly ignored files with erb-formatter
As expected all ERB files starting with the letters 'i', 'c', 'o', 'n'
are formatted.

One thing I'm not keen on is how the ERB formatter converts the ternary
operator into an if-else statement. This makes the code longer and more
difficult to read, and unfortunately, there's no option to disable this
feature.