Enabling to properly cloudify Hetzner host
Updated the cloudify_server script to allow cloudify a Hetzner host as such. This requires passing in a Hetzner server ID in addition to the server IP address. This specifically helps when server has additional Ipv4 subnet added and it needs to be accounted for in our Clover info.
Also added a step to download the default Ubuntu image.
After running this script the clodified host is ready to go.