Move subject part of Authorization.matched_policies query to subquery
Currently, subject is tied to the account, but when we introduce
personal access tokens with more limited access, we'll need to start
performing queries where the subject is a personal access token and
not an account.  This restructures the query to make that future
work simpler.

While here, turn JOIN ... ON TRUE to CROSS JOIN.