Extract tailwind classes of radio_small_cards to a custom class
We puts lots of tailwind classes in the radio_small_cards component. It makes
reading the ERB code harder. In the upcoming commits, we will add branching
where we would need to duplicate the tailwind classes. To make the code more
readable and maintainable, I'm extracting the tailwind classes to a custom
class.