Modify radio_small_cards form component to handle array input
radio_small_cards form component was only handling single input. In the code
there are few cases where we needed to display multiple texts and for those
we had to inline the content of the radio_small_cards and extend instead of
directly using the component. This commit modifies the component to handle
array input and display multiple texts.

Co-authored-by: enes@ubicloud.com