Pop from DestroyGithubInstallation if installation already destroyed
We assemble the DestroyGithubInstallation prog in several places. For
instance, if the same webhook is delivered twice, we end up with
multiple instances for the same installation.

When the first instance destroys the installation, the second one starts
to fail.

We could add a check to ensure only a single instance is running, but
the current implementation is more robust and straightforward.

We already have a similar check in place within the
Prog::Vnet::CertServer prog.