Unify how we trigger failures in the load balancer routes
Generally, we raise `Validation::ValidationFailed` exceptions when we
can't validate the provided body parameters. The load balancer sometimes
uses this too, but not always. Also, when we attempt to retrieve a
record using an ID from the body, it seems more like a BadRequest than a
NotFound to me. I believe NotFound should only be raised when a resource
corresponding to the requested URL is not found, which in this case, is
the load balancer.