Fix two bugs in rake refresh_sequel_caches
Delete the existing cache files before rebuilding them.  Otherwise,
new columns added to existing tables would not be picked up, because
the model would use the cached schema instead of querying the
database.

Always run using the test environment.  This rake task should not
need to be run manually, as it is run automatically by the test_up
and test_down tasks, so it should use the test environment.