Increase nap time at GithubRunner.wait
In the GitHubRunner.wait method, we check whether the runner script is
running or if it has been idle for 5 minutes without picking up a job.
We increment the destroy semaphore upon receiving a webhook event from
GitHub. Previously, I had set the nap time to 15 seconds because I was
unaware that we update the scheduled time when we increase the
semaphore. My intention was to deprovision the virtual machine and free
up our capacity as quickly as possible. However, since we update the
scheduled time when we increase the semaphore, there's no need to run
the wait method frequently. As a result, I've increased the nap time to
60 seconds.

Additionally, this method is one of the most time-consuming for
respirate. I believe this change will help reduce the overall time.