Refactor load balancer post api
We create/retrieve location-based resources using
`project/<P>/location/<L>/vm/<NAME>` pattern [^1]. However, the load
balancer deviates from this pattern. We retrieve the load balancer with
`project/<P>/location/<L>/load-balancer/<NAME>`, but create it with
`project/<P>/load-balancer/<NAME>`. This is because we get the location
from the private subnet's location.  Also `attach-vm`/`detach-vm`
endpoints have the location in the path.

We should align it with other resources to maintain consistency for
location-based resources. We use the location information from the path
to validate the private subnet's location. It can't find private subnet
if the locations don't match.

[^1]: https://www.ubicloud.com/docs/api-reference/overview#creating-resources