Perform same authorization for api and web requests in project list
The project is a special resource, distinct from others. When a user
joins a project, we aim to display it in the project list, regardless of
the user's permissions. The `Project:view` is only applied on the
project detail page. While the web route has the correct query, the api
route does not.