Increase parallel dns query count to 10000
This is set to 150 by default according to dnsmasq manual. When we do
not increase, we may see failures like;
Resolving yum.oracle.com (yum.oracle.com)... failed: Name or service not
known.
wget: unable to resolve host address 'yum.oracle.com'

The theory is that during the setup of runners, many customers install
packages from many different repositories and it is causing dnsmasq to
halt some of these requests.