Final merging of api and web routes
This uses the same route namespace for both api and web routes.

This deletes the stub files previously in routes/api and routes/web,
for routes that were already handled in routes/merged.

For the api-only or web-only routes that were still in routes/api
and routes/web, move to routes/merged, and add appropriate
guards. Unfortunately, review of these files is not easy with git,
as git cannot ignore whitespace changes when doing rename detection.
However, they are generally only indented, with an r.on web? block
(the one api-specific route does not need a check in the route itself).

After the namespace change, Rubocop pointed out that the if/else
branches in the main route had the same code, so this DRYs that
up as well.

This fixes the dashboard and after-login branches in clover.rb to
not accept trailing garbage, or request methods other than GET.
Previously, a web request for /{dashboard,after_login}/some/crap/here
would be treated the same as /{dashboard,after_login}.