Make inspect values use ubids instead of uuids
I think this makes the inspect output more helpful for models that
use ResourceMethods.  We already display the ubid next to the main
model name, but this makes the foreign key values also ubids.

Additionally, since the id is already included in ubid form, there
is no reason to include it in the values, so this removes it from
the values.

This also removes the allocation of many intermediate arrays (2 per
column plus one for the map) for each inspect call.