Proof of concept that access_tag.hyper_tag_table is not needed
As the uuids encode type information, you can determine the table
from the uuid, so the hyper_tag_table column is redundant.

To get the 10-bit type information from the uuid in the database:

* Turn it into a string
* Take characters 16-18 and convert the hex value to integer, modulo 0x400

To get the 10-bit type information from an existing uuid integer:

* Take bits 64-73 (UBID#type_number, though it is commented out)

To get the same 10-bit type information for a class:

* Get base32-encoded 2 byte type prefix from TYPE2CLASSNAME,
  convert to integer (UBID.type_number)

The hyper_tag_table column appears to be used only in a single
place, and this converts that usage to using the hyper_tag_id
value instead.

The same is true of applied_tag.tagged_table, as that is not
even used (it is returned in an authorization query, but all
specs pass when it is removed from the query).