DRY up r.rodauth call in main route
This changes where r.rodauth is called for web routes, moving it
after check_csrf!, rodauth.load_memory, and the root route.

To avoid double checking CSRF token after moving it after
check_csrf!, set check_csrf? false in the Rodauth configuration.

The rodauth.load_memory change is probably better to have before
r.rodauth than after, so if they logged in with the remember cookie,
the rodauth pages will treat them requested as logged in.

For the root route, it just redirects to the login route, so it
doesn't matter if it is before or after r.rodauth.