Remove Clover#api? handling of /api routes
Now that api routes depend purely on the hostname, we don't need
this branch.

Switch from &. to .to_s. to avoid having to write a complicated
test that doesn't provide a Host header in order to keep coverage.
A nil Host header will be turned into the empty string, which will
write in api? returning false.