Introduce new project dashboard
The project dashboard is the first thing users see when they create a
new account. It was designed in the early days and hasn't been updated
since.

We've overlooked the dashboard, despite it being one of the most crucial
pages.

Redesigning it with UX principles in mind is a significant task.

I've just introduced the new project dashboard. I haven't spent a lot of
time on it yet, but we can certainly enhance it over time.

Even though it's a preliminary version, I believe it's an improvement
over the previous one and a good starting point.