WIP: Allocator support for creating a VM in a slice
A new Allocation object computes a cpuset for each candidate host and sets is_valid flag if it can be computed. Once a host is selected, the cpuset is converted to a VmHostSlice object that is created during the Allocator.update phase.

Vm Nexus  updated to account for the slice creation and deletion. Upon creation, VM waits for the slice to be created and enabled. Upon deletion, the slice deletion is triggered.

All this is so far for a 'dedicated' slice, which maps 1:1 to a VM.

Some minimal tests are added. No tests verified yet except the allocator_spec.rb