Full VM creation in a slice
Completing the code changes to create and destroy a VM running in a slice. Updated the logic of VmHostSliceAllocator to wrap around the existing VmHostAllocator and add slice allocation on top of it. The code now correctly takes care of adding/removing core/memory from the host in both slice and non-slice mode and for accounting for used cpu and memory inside the slice when VMs are added and removed. For now only one VM uses a slice, but this logic will stay the same when we add multiple VMs inside a single slice.