Move openapi infrastructure to a directory, get a new openapi snapshot
Some of the linting toolchain does not use hidden files, so they can
be annoying to mix with other top-level directory files.

This also picks up a new version of the openapi.yml from @geemus's
work that is more-correct, and far more complete: note that the length
has doubled.

It's not quite 100% correct, because at least a few hard-to-spec
features were added to the API -- probably accidentally -- that we'd
be better off removing, per daniel@ubicloud.com advisory.

These abnormalities are already identified in a development branch we
can pick through after, that has every single test passing the schema
verification.