Only serialize feature flags in the web serialization
An opaque hash of feature flags is awkward to type in openapi schema,
and was only used in the web views.  So limit it only to a "web"
variant of the serializer, and leave it out of the API.

It raises at least one question: is what's going on for the web views
in fact a "serialization" at all?